Adding integation tests for main.py by command line
