+++ .gitignore
@@ -6,3 +6,4 @@ htmlcov
 dist
 pkg_compare_cv.egg-info
 *.db
+logs
