+++ logs/crawler.log
@@ -7,3 +7,11 @@
 2021-11-24 00:19:47 INFO     Deleting database...
 2021-11-24 00:26:05 INFO     Starting driver.
 2021-11-24 00:27:18 INFO     Starting driver.
+2021-11-24 23:37:55 INFO     Deleting database crawlers...
+2021-11-24 23:38:10 INFO     Deleting database crawlers...
+2021-11-24 23:38:38 INFO     Deleting database crawlers...
+2021-11-24 23:38:55 INFO     Deleting database crawlers...
+2021-11-24 23:39:45 INFO     Deleting database crawlers...
+2021-11-24 23:47:48 INFO     Deleting database crawlers...
+2021-11-24 23:48:00 INFO     Deleting database crawlers...
+2021-11-24 23:48:25 INFO     Deleting database crawlers...
